Implement dynamic social media links, standardized delete endpoints, and optimized API connection settings

+++ src/main/java/com/sarangsvkm/portfolio_api/repository/SkillRepository.java
@@ -5,4 +5,4 @@ package com.sarangsvkm.portfolio_api.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.sarangsvkm.portfolio_api.entity.*;
 
-public interface SkillRepository extends JpaRepository<Skill, Integer> {}
+public interface SkillRepository extends JpaRepository<Skill, Long> {}
